refactor: make State use Double instead of Float

+++ TeamCode/src/main/java/sigmacorns/State.kt
@@ -1,10 +1,10 @@
 package sigmacorns
 
 data class State (
-    var flywheelSpeed: Float,
-    val driveTrainPosition: FloatArray,
-    val driveTrainVelocity: Float,
-    val driveTrainAcceleration: Float,
-    val intakeFlapPosition: Float,
-    val intakeRollerPower: Float
+    var flywheelSpeed: Double,
+    val driveTrainPosition: DoubleArray,
+    val driveTrainVelocity: Double,
+    val driveTrainAcceleration: Double,
+    val intakeFlapPosition: Double,
+    val intakeRollerPower: Double
 )
